+++ runtime/src/python/triton_distributed/runtime/operator.py
@@ -23,7 +23,7 @@ from tritonserver import Server
 
 from triton_distributed.icp.data_plane import DataPlane
 from triton_distributed.icp.request_plane import RequestPlane
-from triton_distributed.worker.remote_request import RemoteInferenceRequest
+from triton_distributed.runtime.remote_request import RemoteInferenceRequest
 
 
 class Operator(abc.ABC):
