+++ runtime/src/python/triton_distributed/runtime/triton_core_operator.py
@@ -26,10 +26,10 @@ from tritonserver import InvalidArgumentError, Server
 
 from triton_distributed.icp.data_plane import DataPlane
 from triton_distributed.icp.request_plane import RequestPlane
-from triton_distributed.worker.logger import get_logger
-from triton_distributed.worker.operator import Operator
-from triton_distributed.worker.remote_request import RemoteInferenceRequest
-from triton_distributed.worker.remote_response import RemoteInferenceResponse
+from triton_distributed.runtime.logger import get_logger
+from triton_distributed.runtime.operator import Operator
+from triton_distributed.runtime.remote_request import RemoteInferenceRequest
+from triton_distributed.runtime.remote_response import RemoteInferenceResponse
 
 
 class TritonCoreOperator(Operator):
